requirements.txt
flask
python-whois
flask_restful
flask_cors
gunicorn
python-dateutil